- Change price type from Int to Double in Product model - Initialize SearchView with viewModel and navigation support - update constraints in ProductTableViewCell for better layout - Enhance search API performance - Begin development of the Product Details Screen interface

+++ Meli/Utils/Extensions/DoubleExtension.swift
@@ -1,5 +1,5 @@
 //
-//  IntExtension.swift
+//  DoubleExtension.swift
 //  Meli
 //
 //  Created by Ricardo Santana on 13/05/24.
@@ -7,7 +7,7 @@
 
 import Foundation
 
-extension Int {
+extension Double {
     func toCurrencyFormat() -> String? {
         let formatter = NumberFormatter()
         formatter.numberStyle = .currency
